+++ btcstaking-tracker/stakingeventwatcher/stakingeventwatcher.go
@@ -973,17 +973,16 @@ func (sew *StakingEventWatcher) fetchStakingTxsByEvent(ctx context.Context, star
 	sew.latency("fetchStakingTxsByEvent")()
 	var stakingTxHashes []string
 
-	criteria := fmt.Sprintf(`tx.height>=%d AND tx.height<=%d AND %s.new_state EXISTS`, startHeight, endHeight, event)
-
-	i := 1
+	page := 1
 	batchSize := 500
 	// #nosec G115 -- performed check
 	if sew.cfg.NewDelegationsBatchSize <= uint64(math.MaxInt) {
 		batchSize = int(sew.cfg.NewDelegationsBatchSize)
 	}
-
+	startHeight = max(startHeight-1, 0)
 	for {
-		stkTxs, err := sew.babylonNodeAdapter.StakingTxHashesByEvent(ctx, event, criteria, &i, &batchSize)
+		criteria := fmt.Sprintf(`tx.height>%d AND tx.height<=%d AND %s.new_state EXISTS`, startHeight, endHeight, event)
+		stkTxs, err := sew.babylonNodeAdapter.StakingTxHashesByEvent(ctx, event, criteria, &page, &batchSize)
 
 		if err != nil {
 			return nil, fmt.Errorf("error fetching active delegations from babylon: %w", err)
@@ -1000,7 +999,7 @@ func (sew *StakingEventWatcher) fetchStakingTxsByEvent(ctx context.Context, star
 			return stakingTxHashes, nil
 		}
 
-		i++
+		page++
 	}
 }