+++ CHANGELOG.md
@@ -41,6 +41,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 * [#357](https://github.com/babylonlabs-io/vigilante/pull/357) imp: handle modules genesis changes
 
+### Bug Fixes
+* [#359](https://github.com/babylonlabs-io/vigilante/pull/359) fix: correctly increment page for stk hash query
+
 ## v0.23.6
 
 ### Improvements