+++ CHANGELOG.md
@@ -37,6 +37,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#376](https://github.com/babylonlabs-io/vigilante/pull/376) chore: bump babylon node dep to v2 rc
+* [#381](https://github.com/babylonlabs-io/vigilante/pull/381) chore: bump babylon node dep to v2
+
 ## v0.23.7
 
 ### Improvements
@@ -47,7 +52,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#362](https://github.com/babylonlabs-io/vigilante/pull/362) chore: manage overlap block range
 * [#369](https://github.com/babylonlabs-io/vigilante/pull/369) chore: do not reprocess block heights - but for real now
 * [#370](https://github.com/babylonlabs-io/vigilante/pull/370) chore: staking event watcher add retry
-* [#376](https://github.com/babylonlabs-io/vigilante/pull/376) chore: bump babylon node dep to v2
 
 ### Bug Fixes
 * [#359](https://github.com/babylonlabs-io/vigilante/pull/359) fix: correctly increment page for stk hash query